# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Helix Protocol API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Email taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── User wallet ───────────────────────────────────────────────────────────
...
  /merchant/transactions:
    get:
      operationId: getMerchantTransactions
      tags: [merchant]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: currency
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Merchant transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserTransaction"

  # ── Payment links ─────────────────────────────────────────────────────────
  /merchant/payment-links:
    get:
      operationId: listPaymentLinks
      tags: [merchant]
      responses:
        "200":
          description: Payment links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentLink"

    post:
      operationId: createPaymentLink
      tags: [merchant]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentLinkRequest"
      responses:
        "201":
          description: Payment link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentLink"
  /merchant/payment-links/{id}:
    delete:
      operationId: cancelPaymentLink
      tags: [merchant]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── POS ───────────────────────────────────────────────────────────────────
  /merchant/pos/charge:
    post:
      operationId: posCharge
      tags: [merchant]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PosChargeRequest"
      responses:
        "200":
          description: Charge successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserTransaction"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Public pay link ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [user, merchant]
        businessName:
          type: string
      required: [email, password, fullName, role]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
        businessName:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, email, fullName, role, createdAt]

    # ── Wallet ──────────────────────────────────────────────────────────────
...
    UserTransaction:
      type: object
      properties:
        id:
          type: string
        senderId:
          type: ["string", "null"]
        recipientId:
          type: ["string", "null"]
        senderAddress:
          type: ["string", "null"]
        recipientAddress:
          type: ["string", "null"]
        amount:
          type: string
        currency:
          type: string
        type:
          type: string
        status:
          type: string
        description:
          type: ["string", "null"]
        chain:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, amount, currency, type, status, createdAt]

    # ── Merchant ────────────────────────────────────────────────────────────
...
    PaymentLink:
      type: object
      properties:
        id:
          type: string
        merchantId:
          type: string
        amount:
          type: string
        currency:
          type: string
        description:
          type: string
        linkCode:
          type: string
        status:
          type: string
        payerId:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
        url:
          type: string
      required: [id, merchantId, amount, currency, description, linkCode, status, createdAt, url]
    CreatePaymentLinkRequest:
      type: object
      properties:
        amount:
          type: number
        currency:
          type: string
        description:
          type: string
        expiresInHours:
          type: integer
      required: [amount, currency, description]

    # ── POS ─────────────────────────────────────────────────────────────────
    PosChargeRequest:
      type: object
      properties:
        amount:
          type: number
        currency:
          type: string
        description:
          type: string
        customerEmail:
          type: string
      required: [amount, currency, description]

    # ── Pay link (public) ───────────────────────────────────────────────────
...