# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Helix Protocol API
 * OpenAPI spec version: 0.2.0
 */
import type { RegisterRequestRole } from "./registerRequestRole";

export interface RegisterRequest {
  email: string;
  password: string;
  fullName: string;
  role: RegisterRequestRole;
  businessName?: string;
}