# lib/api-zod/src/generated/types/userTransaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Helix Protocol API
 * OpenAPI spec version: 0.2.0
 */

export interface UserTransaction {
  id: string;
  /** @nullable */
  senderId?: string | null;
  /** @nullable */
  recipientId?: string | null;
  /** @nullable */
  senderAddress?: string | null;
  /** @nullable */
  recipientAddress?: string | null;
  amount: string;
  currency: string;
  type: string;
  status: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  chain?: string | null;
  createdAt: string;
}